:art: Move packages to install into helpers
