+++ lib/helpers.js
@@ -33,3 +33,22 @@ export function installPackages(dependencies, progressCallback) {
     })
   })
 }
+
+export function packagesToInstall(name) {
+  let packageInfo = atom.packages.getLoadedPackage(name)
+
+  const toInstall = [], toEnable = [];
+  (packageInfo ? (packageInfo.metadata['package-deps'] ? packageInfo.metadata['package-deps'] : []) : [])
+    .forEach(function(name) {
+      if (!window.__steelbrain_package_deps.has(name)) {
+        window.__steelbrain_package_deps.add(name)
+        if (atom.packages.resolvePackagePath(name)) {
+          toEnable.push(name)
+        } else {
+          toInstall.push(name)
+        }
+      }
+    })
+
+  return {toInstall, toEnable}
+}
