+++ lib/main.js
@@ -2,30 +2,11 @@
 const FS = require('fs')
 const Path = require('path')
 const {View} = require('./view')
-import {installPackages} from './install'
+import {installPackages, packagesToInstall} from './helpers'
 
 // Renamed for backward compatibility
 if (typeof window.__steelbrain_package_deps === 'undefined') {
   window.__steelbrain_package_deps = new Set()
-}
-
-export function packagesToInstall(name) {
-  let packageInfo = atom.packages.getLoadedPackage(name)
-
-  const toInstall = [], toEnable = [];
-  (packageInfo ? (packageInfo.metadata['package-deps'] ? packageInfo.metadata['package-deps'] : []) : [])
-    .forEach(function(name) {
-      if (!__steelbrain_package_deps.has(name)) {
-        __steelbrain_package_deps.add(name)
-        if (atom.packages.resolvePackagePath(name)) {
-          toEnable.push(name)
-        } else {
-          toInstall.push(name)
-        }
-      }
-    })
-
-  return {toInstall, toEnable}
 }
 
 export function install(name = null, enablePackages = false) {
